requirements.txt
fastapi==0.115.11
uvicorn==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.6.4
python-dotenv==1.0.1
jinja2==3.1.3
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
anthropic==0.49.0
Pillow==11.1.0
google-generativeai==0.8.4
opencv-python==4.9.0.80
numpy==1.26.4